disassemble mlocationbody
